wip: moved types into separate folders
