+++ backend/src/api-tests/locality/update.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, beforeAll, afterAll, describe, it, expect } from '@jest/globals'
-import { LocalityDetailsType, SpeciesDetailsType } from '../../../../frontend/src/shared/types'
+import { LocalityDetailsType, SpeciesDetailsType } from '../../../../frontend/src/shared/types/dbTypes'
 import { LogRow } from '../../services/write/writeOperations/types'
 import { editedLocality } from './data'
 import { login, resetDatabase, send, testLogRows, resetDatabaseTimeout } from '../utils'
