+++ backend/src/api-tests/utils.ts
@@ -1,7 +1,7 @@
 import request from 'supertest'
 import app from '../app'
 import { expect } from '@jest/globals'
-import { UpdateLog } from '../../../frontend/src/shared/types'
+import { UpdateLog } from '../../../frontend/src/shared/types/dbTypes'
 import { LogRow } from '../services/write/writeOperations/types'
 
 let token: string | null = null
