+++ backend/src/middlewares/authenticator.ts
@@ -2,8 +2,8 @@ import { Middleware } from '../types'
 import { SECRET } from '../utils/config'
 import jwt, { Secret, TokenExpiredError } from 'jsonwebtoken'
 import { nowDb } from '../utils/db'
-import { User } from '../../../frontend/src/shared/types'
-import { Role } from '../../../frontend/src/types'
+import { User } from '../../../frontend/src/shared/types/dbTypes'
+import { Role } from '../../../frontend/src/shared/types'
 
 /*
   This allows extractors to add user and token to request
