+++ backend/src/routes/species.ts
@@ -1,10 +1,10 @@
 import { Request, Router } from 'express'
 import { getAllSpecies, getSpeciesDetails, validateEntireSpecies } from '../services/species'
 import { fixBigInt } from '../utils/common'
-import { EditMetaData, SpeciesDetailsType } from '../../../frontend/src/shared/types'
+import { EditMetaData, SpeciesDetailsType } from '../../../frontend/src/shared/types/dbTypes'
 import { deleteSpecies, writeSpecies } from '../services/write/species'
 import { requireOneOf } from '../middlewares/authorizer'
-import { Role } from '../../../frontend/src/types'
+import { Role } from '../../../frontend/src/shared/types'
 
 const router = Router()
