+++ backend/src/services/localitySpeciesExport.ts
@@ -1,7 +1,7 @@
 import { PoolConnection } from 'mariadb'
-import { User } from '../../../frontend/src/shared/types'
+import { User } from '../../../frontend/src/shared/types/dbTypes'
 import { NOW_DB_NAME } from '../utils/config'
-import { Role } from '../../../frontend/src/types'
+import { Role } from '../../../frontend/src/shared/types'
 import { pool } from '../utils/db'
 
 const getExportList = async (
