+++ backend/src/services/write/locality.ts
@@ -1,8 +1,8 @@
-import { EditDataType, LocalityDetailsType, Reference, User } from '../../../../frontend/src/shared/types'
+import { EditDataType, LocalityDetailsType, Reference, User } from '../../../../frontend/src/shared/types/dbTypes'
 import { NOW_DB_NAME } from '../../utils/config'
 import { WriteHandler } from './writeOperations/writeHandler'
 import { getFieldsOfTables } from '../../utils/db'
-import { getHomininSkeletalRemains } from '../../../../frontend/src/types'
+import { getHomininSkeletalRemains } from '../../../../frontend/src/shared/types'
 import { getLocalityDetails, filterDuplicateLocalitySpecies } from '../locality'
 import { ActionType } from './writeOperations/types'
 import { makeListRemoved, fixRadioSelection } from './writeOperations/utils'
