+++ backend/src/services/write/timeBound.ts
@@ -5,7 +5,7 @@ import {
   LocalityDetailsType,
   TimeUnitDetailsType,
   User,
-} from '../../../../frontend/src/shared/types'
+} from '../../../../frontend/src/shared/types/dbTypes'
 import { NOW_DB_NAME } from '../../utils/config'
 import { WriteHandler } from './writeOperations/writeHandler'
 import { getFieldsOfTables } from '../../utils/db'
