+++ backend/src/services/write/writeOperations/writeHandler.ts
@@ -1,4 +1,4 @@
-import { Reference, RowState } from '../../../../../frontend/src/shared/types'
+import { Reference, RowState } from '../../../../../frontend/src/shared/types/dbTypes'
 import { logger } from '../../../utils/logger'
 import { logAllUpdates } from './updateLogger/prepareUpdates'
 import { DatabaseHandler, DbWriteItem } from './databaseHandler'
