+++ frontend/src/shared/calculations.ts
@@ -1,14 +1,5 @@
-export enum Role {
-  Admin = 1,
-  EditUnrestricted = 2,
-  EditRestricted = 3,
-  NowOffice = 4,
-  Project = 5,
-  ProjectPrivate = 6,
-  ReadOnly = 7,
-}
 
-import { EditDataType, LocalityDetailsType } from '../src/shared/types'
+import { EditDataType, LocalityDetailsType } from './types'
 
 export const calculateMeanHypsodonty = (locality: LocalityDetailsType) => {
   // Mean hypsodonty calculation has been made based on html/include/database.php line 2567 onwards
