+++ frontend/src/shared/types/misc.ts
@@ -1,0 +1,27 @@
+export enum Role {
+    Admin = 1,
+    EditUnrestricted = 2,
+    EditRestricted = 3,
+    NowOffice = 4,
+    Project = 5,
+    ProjectPrivate = 6,
+    ReadOnly = 7,
+  }
+
+export type User = {
+    username: string
+    role: Role
+    userId: number
+    initials: string
+  }
+  
+export type ValidationErrors = {
+    status: string
+    data: ValidationErrorItem[]
+  }
+  
+  type ValidationErrorItem = {
+    name: string
+    error: string
+  }
+
